Split out properties controller from things controller
