+++ src/controllers/things_controller.ts
@@ -8,6 +8,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
+import PropertiesController from './properties_controller';
 import Action, { ActionDescription } from '../models/action';
 import Actions from '../models/actions';
 import ActionsController from './actions_controller';
@@ -25,7 +26,6 @@ import { WithJWT } from '../jwt-middleware';
 import { Any } from 'gateway-addon/lib/schema';
 import { Property } from 'gateway-addon';
 import { isW3CThingDescription } from '../utils';
-import { HttpErrorWithCode } from '../errors';
 
 interface SetPropertyMessage {
   messageType: 'setProperty';
@@ -293,94 +293,9 @@ function build(): express.Router {
   });
 
   /**
-   * Get the properties of a Thing.
+   * Use a PropertiesController to handle each thing's properties.
    */
-  controller.get('/:thingId/properties', async (request, response) => {
-    const thingId = request.params.thingId;
-
-    let thing;
-    try {
-      thing = await Things.getThing(thingId);
-    } catch (e) {
-      console.error('Failed to get thing:', e);
-      response.status(404).send(e);
-      return;
-    }
-
-    const result: Record<string, Any> = {};
-    for (const name in thing.getProperties()) {
-      try {
-        const value = await AddonManager.getProperty(thingId, name);
-        result[name] = value;
-      } catch (e) {
-        console.error(`Failed to get property ${name}:`, e);
-      }
-    }
-
-    response.status(200).json(result);
-  });
-
-  /**
-   * Set multiple properties of a Thing.
-   */
-  controller.put('/:thingId/properties', async (request, response) => {
-    const thingId = request.params.thingId;
-    if (!(typeof request.body === 'object') || request.body === null) {
-      response.sendStatus(400);
-      return;
-    }
-    // An array of promises to set each property
-    const promises = [];
-    for (const propertyName of Object.keys(request.body)) {
-      promises.push(Things.setThingProperty(thingId, propertyName, request.body[propertyName]));
-    }
-    Promise.all(promises)
-      .then(() => {
-        // Respond with success code if all properties set successfully
-        response.sendStatus(204);
-      })
-      .catch((err) => {
-        // Otherwise send an error response
-        console.error('Error setting property:', err);
-        response.sendStatus(500);
-      });
-  });
-
-  /**
-   * Get a property of a Thing.
-   */
-  controller.get('/:thingId/properties/:propertyName', async (request, response) => {
-    const thingId = request.params.thingId;
-    const propertyName = request.params.propertyName;
-    try {
-      const value = await Things.getThingProperty(thingId, propertyName);
-      response.status(200).json(value);
-    } catch (err) {
-      response.status((err as HttpErrorWithCode).code).send((err as HttpErrorWithCode).message);
-    }
-  });
-
-  /**
-   * Set a property of a Thing.
-   */
-  controller.put('/:thingId/properties/:propertyName', async (request, response) => {
-    const thingId = request.params.thingId;
-    const propertyName = request.params.propertyName;
-    if (typeof request.body === 'undefined') {
-      response.sendStatus(400);
-      return;
-    }
-    const value = request.body;
-    try {
-      await Things.setThingProperty(thingId, propertyName, value);
-      response.sendStatus(204);
-    } catch (err) {
-      console.error('Error setting property:', err);
-      response
-        .status((err as HttpErrorWithCode).code || 500)
-        .send((err as HttpErrorWithCode).message);
-    }
-  });
+  controller.use(`/:thingId${Constants.PROPERTIES_PATH}`, PropertiesController());
 
   /**
    * Use an ActionsController to handle each thing's actions.
